Update visual appearance of the robots handler and update 3rd party JS References.
