+++ src/Stott.Optimizely.RobotsHandler/Presentation/ViewModels/RobotsListItemViewModel.cs
@@ -1,0 +1,12 @@
+﻿using System;
+
+namespace Stott.Optimizely.RobotsHandler.Presentation.ViewModels;
+
+public class RobotsListItemViewModel
+{
+    public Guid Id { get; set; }
+
+    public string Name { get; set; }
+
+    public string Url { get; set; }
+}
