+++ src/Stott.Optimizely.RobotsHandler/Views/Robots/RobotsSiteList.cshtml
@@ -1,42 +1,48 @@
-﻿@model Stott.Optimizely.RobotsHandler.Presentation.ViewModels.RobotsListViewModel
+﻿@using EPiServer.Shell.Navigation
+@model Stott.Optimizely.RobotsHandler.Presentation.ViewModels.RobotsListViewModel
 
 @{
-    Layout = "_RobotsLayout";
-    ViewData["Title"] = "Robots.txt Management";
+    Layout = "";
 }
 
-<div class="row">
-    <div class="col">
-        <h1 class="js-header">
-            @ViewData["Title"]
-        </h1>
-
-        <table class="table table-striped">
-            <thead>
-                <tr>
-                    <th>Site Name</th>
-                    <th>Site Url</th>
-                    <th>Actions</th>
-                </tr>
-            </thead>
-            <tbody>
-                @foreach (var site in Model.List)
-                {
+<!DOCTYPE html>
+<html lang="en">
+<head>
+    <title>@Model.ApplicationName</title>
+    <link href="https://cdn.jsdelivr.net/npm/bootstrap@5.2.3/dist/css/bootstrap.min.css" rel="stylesheet" integrity="sha384-rbsA2VBKQhggwzxH7pPCaAqO46MgnOM80zW1RWuH61DGLwZJEdK2Kadq2F9CUG65" crossorigin="anonymous">
+</head>
+<body>
+    @Html.CreatePlatformNavigationMenu()
+    <div @Html.ApplyPlatformNavigation()>
+        <div class="container-fluid p-2 bg-dark text-light">
+            <p class="my-0">
+                <span class="h5">@Model.ApplicationName</span><span class="mx-3">@Model.ApplicationVersion</span>
+            </p>
+        </div>
+        <div class="container-xxl py-3" style="border-top: 1px solid #d6d6d6;">
+            <table class="table table-striped">
+                <thead>
                     <tr>
-                        <td>@site.Name</td>
-                        <td>@site.Url</td>
-                        <td>
-                            <button class="btn btn-primary js-edit-button" data-siteid="@site.Id">Edit</button>
-                        </td>
+                        <th>Site Name</th>
+                        <th>Site Url</th>
+                        <th>Actions</th>
                     </tr>
-                }
-            </tbody>
-        </table>
-
+                </thead>
+                <tbody>
+                    @foreach (var site in Model.List)
+                    {
+                        <tr>
+                            <td>@site.Name</td>
+                            <td>@site.Url</td>
+                            <td>
+                                <button class="btn btn-primary js-edit-button" data-siteid="@site.Id">Edit</button>
+                            </td>
+                        </tr>
+                    }
+                </tbody>
+            </table>
+        </div>
     </div>
-</div>
-
-@section modals {
 
     <div id="edit-robots-content-modal" class="modal" tabindex="-1">
         <div class="modal-dialog">
@@ -65,58 +71,59 @@
         </div>
     </div>
 
-}
+    <script src="https://cdn.jsdelivr.net/npm/bootstrap@5.2.3/dist/js/bootstrap.bundle.min.js" integrity="sha384-kenU1KFdBIe4zVF0s0G1M5b4hcpxyD9F7jL+jjXkk+Q2h455rYXK/7HAuoJl+0I4" crossorigin="anonymous"></script>
+    <script src="https://cdn.jsdelivr.net/npm/jquery@3.6.4/dist/jquery.min.js" integrity="sha256-oP6HI9z1XaZNBrJURtCoUT5SUnxFr8s3BzRl+cbzUq8=" crossorigin="anonymous"></script>
 
-@section scripts {
     <script>
-        $(document).ready(function() {
+        $(document).ready(function () {
 
             var myModal = new bootstrap.Modal(document.getElementById('edit-robots-content-modal'), {
                 keyboard: false
             });
 
-            $('.js-edit-button').click(function() {
+            $('.js-edit-button').click(function () {
                 let siteId = $(this).data('siteid');
                 $.get('/Robots/Details/', { siteId: siteId })
-                    .done(function(data) {
+                    .done(function (data) {
                         $('.js-modal-title').html(data.siteName);
                         $('.js-modal-siteid').html(data.siteId);
                         $('.js-modal-robots-content').val(data.robotsContent);
 
                         myModal.toggle();
                     })
-                    .fail(function() {
+                    .fail(function () {
                         let loadError = '<div class="alert alert-danger alert-dismissible fade show" role="alert">'
-                                      + '<strong>Failure!</strong> An error was encountered when trying to retrieve your robots.txt content.'
-                                      + '<button type="button" class="btn-close" data-bs-dismiss="alert" aria-label="Close"></button>'
-                                      + '</div>';
+                            + '<strong>Failure!</strong> An error was encountered when trying to retrieve your robots.txt content.'
+                            + '<button type="button" class="btn-close" data-bs-dismiss="alert" aria-label="Close"></button>'
+                            + '</div>';
                         $(loadError).insertAfter('.js-header');
                     });
             });
 
-            $('.js-save-button').click(function() {
+            $('.js-save-button').click(function () {
                 let siteName = $('.js-modal-title').text();
                 let siteId = $('.js-modal-siteid').text();
                 let robotsContent = $('.js-modal-robots-content').val();
 
                 $.post('/Robots/Save', { siteId: siteId, siteName: siteName, robotsContent: robotsContent })
-                    .done(function() {
+                    .done(function () {
                         let saveSuccess = '<div class="alert alert-success alert-dismissible fade show" role="alert">'
-                                        + '<strong>Success!</strong> Your robots.txt content changes were successfully applied.'
-                                        + '<button type="button" class="btn-close" data-bs-dismiss="alert" aria-label="Close"></button>'
-                                        + '</div>';
+                            + '<strong>Success!</strong> Your robots.txt content changes were successfully applied.'
+                            + '<button type="button" class="btn-close" data-bs-dismiss="alert" aria-label="Close"></button>'
+                            + '</div>';
                         $(saveSuccess).insertAfter('.js-header');
                         myModal.toggle();
                     })
-                    .fail(function() {
+                    .fail(function () {
                         let saveError = '<div class="alert alert-danger alert-dismissible fade show" role="alert">'
-                                      + '<strong>Failure!</strong> An error was encountered when trying to save your robots.txt content.'
-                                      + '<button type="button" class="btn-close" data-bs-dismiss="alert" aria-label="Close"></button>'
-                                      + '</div>';
+                            + '<strong>Failure!</strong> An error was encountered when trying to save your robots.txt content.'
+                            + '<button type="button" class="btn-close" data-bs-dismiss="alert" aria-label="Close"></button>'
+                            + '</div>';
                         $(saveError).insertAfter('.js-header');
                         myModal.toggle();
                     });
             });
         });
     </script>
-}
+</body>
+</html>
